Enable detecting SyncTERM DA1 response.
SyncTERM sends CSI = Pn c not CSI ? Pn c like VT terminals.
This results in notcurses hanging in init.

This required WezTerm da3 support to be removed, because the SyncTERM
and WezTerm responses could not both exist in the automaton.  As it
turns out, a dig through the WezTerm git history shows that it never
actually sent CSI = ... ST in response to DA3, and notcurses took
no actions in response to it, so it should be safe to remove this
feature.