sexblitter: fix inverted characters 32+64
while working on the octblitter (see #2669), i noticed that
the last character in the sex[] collection was inverted.
this would result in a swapped foreground/background
color when either the 32nd or 64th spot was hit.
we wanted U+1fb2e (🬮, BLOCK SEXTANT-156) but were
supplying U+1fb0d (🬍, BLOCK SEXTANT-234).