ncplane_put: don't run tabs over the end of plane
as reported by kmarius in
https://github.com/dankamongmen/notcurses/issues/2870
we coredumped if the plane was not a multiple of the tab
stop, and we ended the plane with a tab character. check
for the plane's x boundary while expanding a tab, and
break out of the expansion loop if we hit it. potential
autogrowth of the plane has already been handled above.

closes #2870